requirements.txt
absl-py==1.4.0
aiohttp
aiosignal
appdirs==1.4.4
asynctest==0.13.0
async-timeout
attrs
backcall
blinker
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi
charset-normalizer==3.0.1
charset-normalizer
click
colorama
cryptography
cycler
debugpy
decorator
docker-pycreds
entrypoints
flit_core
frozenlist
gitdb
GitPython
grpcio==1.51.1
grpcio
idna==3.4
idna
imageio==2.25.0
importlib-metadata==6.0.0
importlib-metadata
ipykernel
ipython
jedi
joblib==1.2.0
jupyter_client
jupyter_core
kiwisolver
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib
matplotlib-inline
mkl-fft
mkl-random
mkl-service==2.4.0
multidict
nest-asyncio
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11
oauthlib
olefile
packaging
parso
pathtools
pexpect
pickleshare
Pillow==9.4.0
prompt-toolkit
protobuf==3.20.3
psutil
ptyprocess
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
Pygments
PyJWT
pyOpenSSL
pyparsing
PySocks
python-dateutil
pyu2f
PyYAML
pyzmq
requests==2.28.2
requests-oauthlib==1.3.1
rsa
scikit-learn
scipy
setproctitle
six==1.16.0
smmap
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado
tqdm
traitlets
typing_extensions==4.4.0
urllib3==1.26.14
wandb
wcwidth
Werkzeug==2.2.2
wget==3.2
yarl
zipp==3.12.1
